Create Fibonacci series
The fibonacci series starts from 0, 1,(as there is no (n-1),(n-2) for the numbers 0 or 1).
Hence, if the user input is 0 or 1 we return the same number. 
Storing the fibonacci sequence in the 'fib_series' variable to calculate the series 
Looping through from (2 - user input + 1).
Now, appending the value of ([user_input  -  1] + [user_input  -  2]) to the series and returning it.